packages: add builds for netfilter packages from upstream Debian
Instead of scattering build instructions accross individual repositories for
some additional packages we want to build from latest Debian releases, this
is a proof-of-concept how to integrate an individual package Pipeline into the
overall VyOS build repository.

A dedicated Jenkins job will be required for this Pipeline but it will only
compile if files are actually changes in the directory path relevant for
this components.