Kernel: T3489: re-enable NUMA in kernel build
All throughout the history of vyos-build, NUMA has been disabled.
It WAS enabled back in Vyatta days, so something somewhere has
accidentally turned it off - and it's ALSO mentioned explicitly
in various pieces of documentation, so it's MEANT to be on!

Note for future researchers: CONFIG_INOTIFY_STACKFS=y is provided
by a patch, so patch the kernel BEFORE running make menuconfig, or
manually put it back. If not, it'll vanish and is a wanted part of
the build.

Signed-Off-By: Rob Thomas <xrobau@gmail.com>