Kernel: T3587: replace Out-Of-Tree QAT drivers with In-Tree drivers
The latest Intel QAT drivers make the 5.10 series Kernel crash and there was
no real and good response from Intel so far.

Today the build failed one more time by a source file moving around in the
Intel webservers. It's time to try to move the QAT functionality to the Kernel
buildin drivers.