T4966: Delay UDEV execution, avoid rename deadlock
UDEV will rename interfaces from whatever the kernel called them
to eX before converting them to ethX during init. In current VyOS,
the second renaming operation can run into a lock on the adapter
preventing altering its name. As a result, the adapter will remain
in the eX configuration, preventing proper execution of subsequent
scripts and configuration stanzas.

The initial renaming step has to remain as it is needed to work
around other issues, which leaves the somewhat hacky approach of
delaying the second renaming step slightly in an effort to let the
device lock holders settle, releasing it for rename to ethX. This
is accomplished by a kernel commandline paramter (3s), which can be
tweaked to reduce impact or wait longer as needed on different
devices - udev.exec_delay=3