frr: T6023: add pending upstream patches
Latest stable/9.0 FRR branch has an updated version of frr-reload.py which
is more strict on the commands to remove from the running config. This resulted
in some issues in the codebase of FRR which are addressed in the attached patches
and are pending for upstream integration.

(cherry picked from commit 798b95eb9776486aed874346869a8238383c18f6)